add _xyz2mol.py to shorten molconvert
